refactor: service/usecase を completion/ サブディレクトリに整理
将来的に completion 以外の機能（analytics など）を追加しやすくするため、
service と usecase の配下に completion/ サブディレクトリを作成し、
既存のファイルを移動。

- llm_chat/domain/service/completion/
  - base.py, common.py, riddle.py, multimedia.py, rag.py
- llm_chat/domain/usecase/completion/
  - base.py, common.py, riddle.py, multimedia.py, rag.py

各ディレクトリの __init__.py を更新し、後方互換性を維持。
既存のテストがすべてパス (12 tests OK)
