+++ llm_chat/domain/service/completion/__init__.py
@@ -1,17 +1,21 @@
-from llm_chat.domain.service.completion import (
-    BaseChatService,
+from llm_chat.domain.service.completion.base import BaseChatService
+from llm_chat.domain.service.completion.common import (
     ChatService,
     OpenAIChatStreamingService,
     get_chat_history,
+)
+from llm_chat.domain.service.completion.riddle import (
     RiddleTask,
     RIDDLE_END_MESSAGE,
     get_prompt,
     create_initial_prompt,
+)
+from llm_chat.domain.service.completion.multimedia import (
     OpenAIDalleChatService,
     OpenAITextToSpeechChatService,
     OpenAISpeechToTextChatService,
-    OpenAIRagChatService,
 )
+from llm_chat.domain.service.completion.rag import OpenAIRagChatService
 
 __all__ = [
     "BaseChatService",
