+++ llm_chat/domain/service/completion/base.py
