+++ llm_chat/domain/service/completion/riddle.py
