+++ llm_chat/domain/usecase/completion/base.py
