+++ llm_chat/domain/usecase/completion/common.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import User
 from lib.llm.service.completion import LlmCompletionStreamingService
 from lib.llm.valueobject.completion import RoleType, StreamResponse
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig
-from llm_chat.domain.service.common import ChatService, OpenAIChatStreamingService
-from llm_chat.domain.usecase.base import UseCase
+from llm_chat.domain.service.completion.common import ChatService, OpenAIChatStreamingService
+from llm_chat.domain.usecase.completion.base import UseCase
 from llm_chat.domain.valueobject.chat import MessageDTO
