+++ llm_chat/domain/usecase/completion/riddle.py
@@ -2,9 +2,9 @@ from django.contrib.auth.models import User
 
 from lib.llm.valueobject.completion import RoleType
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig
-from llm_chat.domain.service.common import ChatService
-from llm_chat.domain.service.riddle import RIDDLE_END_MESSAGE
-from llm_chat.domain.usecase.base import UseCase
+from llm_chat.domain.service.completion.common import ChatService
+from llm_chat.domain.service.completion.riddle import RIDDLE_END_MESSAGE
+from llm_chat.domain.usecase.completion.base import UseCase
 from llm_chat.domain.valueobject.chat import MessageDTO, GenderType, Gender
